feat: update tooltip to fc
